Refactor Middleware for Scalability and Security Enhancements
### Changes:
#### **Backend (`index.ts`)**:
1. **Middleware Modularization**:
   - Replaced inline CSRF and session handling with modularized middleware:
     - `csrfMiddleware`: Handles CSRF protection.
     - `secFetchSiteMiddleware`: Enforces `Sec-Fetch-Site` security headers.
     - `sessionMiddleware`: Validates sessions using JWT from cookies.

2. **Optimized Middleware Usage**:
   - Applied `csrfMiddleware` and `secFetchSiteMiddleware` globally with `app.use('*')`.
   - Applied `sessionMiddleware` to `/api/*` routes for secure API interactions.

3. **Improved Routing**:
   - Retained modular `translate` and `turnstile` route definitions for clarity and maintainability.

#### **Session Management (`session.ts`)**:
1. **Enhanced `sessionMiddleware`**:
   - Created a factory-based middleware for session validation using JWT and cookie support.
   - Encapsulated logic in `createMiddleware` for reusable, clear implementation.

2. **Refined `createSessionCookie`**:
   - Centralized cookie name management with the constant `COOKIE_NAME`.
   - Maintained strict cookie security settings:
     - `httpOnly`: Protects against client-side JavaScript access.
     - `secure`: Ensures cookies are only transmitted over HTTPS.
     - `sameSite: 'Strict'`: Prevents cross-site request usage.

### Purpose:
This update improves scalability, maintainability, and security by:
- Modularizing middleware for clear separation of concerns.
- Ensuring robust CSRF and session protection for secure user interactions.
- Streamlining session management with reusable patterns and secure defaults.
