+++ src/index.ts
@@ -1,30 +1,16 @@
 import { Hono } from 'hono'
 import translate from './translate'
 import turnstile from './turnstile'
-import { jwt } from 'hono/jwt'
-import { csrf } from 'hono/csrf'
+import { sessionMiddleware } from './session'
+import { csrfMiddleware, secFetchSiteMiddleware } from './csrf'
 
-type Bindings = {
-  JWT_SECRET: string,
-  ALLOWED_ORIGIN: string,
-}
+const app = new Hono()
 
-const app = new Hono<{ Bindings: Bindings }>()
+app.use('*', csrfMiddleware())
+app.use('*', secFetchSiteMiddleware())
+app.use('/api/*', sessionMiddleware())
 
-app.use('*', (c, next) => {
-  const csrfMiddleware = csrf({ origin: c.env.ALLOWED_ORIGIN })
-  return csrfMiddleware(c, next)
-})
-
-app.use('/api/*', (c, next) => {
-  const jwtMiddleware = jwt({
-    secret: c.env.JWT_SECRET,
-    cookie: 'session',
-  })
-  return jwtMiddleware(c, next)
-})
-
-app.route('/api/translate', translate)
 app.route('/auth', turnstile)
+app.route('/api/translate', translate)
 
 export default app
